refactor base code
