+++ homeassistant/components/roborock/manifest.json
@@ -7,7 +7,7 @@
   "iot_class": "local_polling",
   "loggers": ["roborock"],
   "requirements": [
-    "python-roborock==0.40.0",
+    "python-roborock==1.0.0",
     "vacuum-map-parser-roborock==0.1.1"
   ]
 }
