+++ requirements_all.txt
@@ -2297,7 +2297,7 @@ python-rabbitair==0.0.8
 python-ripple-api==0.0.3
 
 # homeassistant.components.roborock
-python-roborock==0.40.0
+python-roborock==1.0.0
 
 # homeassistant.components.smarttub
 python-smarttub==0.0.36
