+++ requirements_test_all.txt
@@ -1773,7 +1773,7 @@ python-qbittorrent==0.4.3
 python-rabbitair==0.0.8
 
 # homeassistant.components.roborock
-python-roborock==0.40.0
+python-roborock==1.0.0
 
 # homeassistant.components.smarttub
 python-smarttub==0.0.36
